Refactor to move converter functions to separate file.
